refactor: staticlize analyst
