+++ src/core/index.ts
@@ -5,6 +5,7 @@ export * from './KeyDetector'
 export * from './Config'
 export * from './CurrentFile'
 export * from './Translator'
+export * from './Analyst'
 export * from './loaders/Loader'
 export * from './loaders/SfcLoader'
 export * from './loaders/LocaleLoader'
